cmd/view-fits: improved performances and features
This CL improves the performances of the images' display.
Previously, an image would be redisplayed whenever a paint.Event would
be received.
This is wasteful, especially when the window is being manually resized
or adjusted to the image dimensions.

Instead, only redraw when the user explicitly asks for it (by pressing
the 'r' key) or when the image window gains focus
(lifecycle.StageVisible->lifecycle.StageFocused.)

Also, adds the following keybindings:
- r: reload
- z: resize/adjust (not yet functional)
- ?: display help
- p: print/save current image into/as output.png