fitsio: reduce memory pressure in readBlock
This CL modifies the internal readBlock function to only allocate when
needed.
readBlock is now given an input, optional, byte slice to be populated
with the extracted HDU block.

As readBlock is only used in one sequential function (DecodeHDU) we can
reuse that block data buffer from one block to the other.

Fixes astrogo/fitsio#31.