add ability to write u16 directly
The storage of int16 and uint16 is identical for FITS; signed integers
are binary two's complement which are bitwise identical to a shifted
uint16.  Therefore, the caller can write with u16, which is faster
since it avoids the conversion from i16=>u16, or indeed two conversions
if they changed their data to i16 from u16 to begin with.